requirements-tests.txt
fabric
mox>=0.5.3
django-test-utilities>=0.5
model-mommy==1.2.6
nose
django-nose
south
